#246 Error en esborrar un procediment amb notificacions
A més d'arreglar aquest error, també s'ha fet que no doni error eliminar
un procediment amb permisos ni grups. O sigui que per al moment
d'eliminar el procediment primer es comprova que no tengui
notificacions, si en té no es permet eliminar, però si no en té
automaticament elimina els grups i el permisos que pugui tenir.

+++ notib-war/src/main/java/es/caib/notib/war/controller/ProcedimentController.java
@@ -213,16 +213,31 @@ public class ProcedimentController extends BaseUserController{
 			HttpServletRequest request,
 			@PathVariable Long procedimentId) {		
 		
-		EntitatDto entitat = getEntitatActualComprovantPermisos(request);
-		
-		procedimentService.delete(
-				entitat.getId(),
-				procedimentId);
-		
-		return getAjaxControllerReturnValueSuccess(
-				request,
-				"redirect:../../procediment",
-				"procediment.controller.esborrat.ok");
+		try {
+			EntitatDto entitat = getEntitatActualComprovantPermisos(request);
+			
+			if (procedimentService.procedimentEnUs(procedimentId)) {
+				return getAjaxControllerReturnValueError(
+						request,
+						"redirect:../../procediment",
+						"procediment.controller.esborrat.enUs");
+			} else {
+				procedimentService.delete(
+						entitat.getId(),
+						procedimentId);
+				
+				return getAjaxControllerReturnValueSuccess(
+						request,
+						"redirect:../../procediment",
+						"procediment.controller.esborrat.ok");
+			}
+		} catch (Exception e) {
+			return getAjaxControllerReturnValueError(
+					request,
+					"redirect:../../procediment",
+					"procediment.controller.esborrat.ko",
+					e);
+		}
 	}
 	
 	@RequestMapping(value = "/update/auto", method = RequestMethod.GET)
